Move module level functions to top. Refactor them.
There is now a single function for getting a `Map` from a `dist_type`. This
pattern was duplicated a lot and is now consolidated in `choose_map`.

Renamed `dist` -> `dist_type` in some places for consitency.

Renamed `map_factor` -> `map_from_sizes` to be consisten with
`map_from_dim_datas` and `map_from_global_dim_dict`.